New networking + Fixes + updates
Note: There might be bugs here and there.

+++ server/Core/World/CoreUpdate.cs
@@ -35,7 +35,7 @@ namespace LoESoft.Server.Core.World
 
                 while (_canpdate)
                 {
-                    _worldTime.TotalElapsedMs = (int)_watch.Elapsed.TotalMilliseconds;
+                    _worldTime.TotalElapsedMs = _watch.Elapsed.TotalMilliseconds;
                     _worldTime.TickCount = tickCount;
 
                     Map.Update(_worldTime);
